requirements.txt
torch==2.6.0
torchvision~=0.21.0
faiss-cpu==1.10.0
huggingface-hub==0.32.0
git+https://github.com/huggingface/transformers.git@6e3063422c4b1c014aa60c32b9254fd2902f0f28
Pillow==10.1.0
numpy==1.26.2
matplotlib==3.8.2